moves import around again, asf-search now imports again
