+++ asf_search/CMR/__init__.py
@@ -3,4 +3,4 @@ from .subquery import build_subqueries
 from .translate import translate_product, translate_opts, get_additional_fields
 from .field_map import field_map
 from .UMMFields import umm_property_paths, umm_property_typecasting
-from .datasets import dataset_collections, datset_product_types
+from .datasets import dataset_collections
