+++ asf_search/__init__.py
@@ -23,6 +23,7 @@ from .ASFSession import ASFSession
 from .ASFProduct import ASFProduct
 from .ASFSearchResults import ASFSearchResults
 from .ASFSearchOptions import ASFSearchOptions, validators
+from .Products import *
 from .exceptions import *
 from .constants import *
 from .health import *
